added list, sync button
