+++ api/mod/util.js
@@ -1,39 +1,16 @@
-const cal = `BEGIN:VCALENDAR
-VERSION:2.0
-CALSCALE:GREGORIAN
-BEGIN:VEVENT
-SUMMARY:Access-A-Ride Pickup
-DTSTART;TZID=America/New_York:20230222T103400
-DTEND;TZID=America/New_York:20230222T110400
-LOCATION:1000 Broadway Ave.\, Brooklyn
-DESCRIPTION: Access-A-Ride trip to 900 Jay St.\, Brooklyn
-STATUS:CONFIRMED
-SEQUENCE:3
-BEGIN:VALARM
-TRIGGER:-PT10M
-DESCRIPTION:Pickup Reminder
-ACTION:DISPLAY
-END:VALARM
-END:VEVENT
-BEGIN:VEVENT
-SUMMARY:Access-A-Ride Pickup
-DTSTART;TZID=America/New_York:20230222T200000
-DTEND;TZID=America/New_York:20230222T203000
-LOCATION:900 Jay St.\, Brooklyn
-DESCRIPTION: Access-A-Ride trip to 1000 Broadway Ave.\, Brooklyn
-STATUS:CONFIRMED
-SEQUENCE:3
-BEGIN:VALARM
-TRIGGER:-PT10M
-DESCRIPTION:Pickup Reminder
-ACTION:DISPLAY
-END:VALARM
-END:VEVENT
-END:VCALENDAR
-`
+const ACCEPT = 'accept'
 module.exports = {
 	setup(host, cfg, rsc, paths){
 		return this
+	},
+
+	branchByContentType(map, def){
+		return function(req){
+			const ct = req.headers[ACCEPT]
+			const route = map[ct] || map[def]
+			if (!route) return this.next(`no route for ${ACCEPT}: ${ct}`)
+			return this.next(null, route)
+		}
 	},
 
 	async wait(sec){
@@ -42,10 +19,4 @@ module.exports = {
 		})
 		return this.next()
 	},
-
-	getCal(key, output){
-	console.log('####', output)
-		output[key] = cal
-		return this.next()
-	}
 }
