+++ ui/mod/ListView.js
@@ -1,0 +1,11 @@
+return {
+	deps: {
+		list: 'models',
+		Item: 'View'
+	},
+	create(deps, params){
+		deps.list.forEach((model, i) => {
+			this.spawn(deps.Item, {i})
+		})
+	}
+}
