+++ ui/mod/events/ButtonSync.js
@@ -1,0 +1,6 @@
+return {
+	deps: {
+	},
+	create(deps, params){
+	}
+}
