+++ ui/mod/events/View.js
@@ -1,0 +1,9 @@
+return {
+	deps: {
+		tmpl: 'file',
+		model: 'model'	
+	},
+	create(deps, params){
+		this.el = deps.tmpl({d:deps.model})
+	}
+}
