+++ ui/mod/events/tmpl.asp
@@ -1,0 +1,5 @@
+<span><%d.summary%></span>
+<span><%d.description%></span>
+<span><%d.location.title%></span>
+<span><%d.start%></span>
+<span><%d.duration%> Mins</span>
